Provide an option to keep attributes for Modules and Functions (#124)
* Record attributes to modules or torch functions

* Stringify attributes

When generating the graph, keep the attributes for for Module and
Function nodes.

Add a test for two simple examples.

* Limit number of attributes saved

* Minor cleanup

* Revert graph name in test, not needed now

* Clean up stringify_attributes function

MyPy clean-up and unused arguments
Method always called by RecorderTensor
Include kwargs for FunctionNode

* Collect attributes enabled by a flag

collect_attributes flag added to draw_graph
If attributes are not collected they will be None

* Add future annotations to test_attributes